AzureSQLPromptFlowSamples/src/requirements.txt
azure-ai-ml
azure-core
azure-identity
azure-keyvault
azure-kusto-data
azure-search-documents>=11.4.0b11
azure-storage-blob
pyodbc
bs4
jinja2
jsbeautifier
langchain
logger
matplotlib
openai
pandas
plotly
promptflow>=0.1.0b8
promptflow-tools
python-dotenv
ipywidgets
scikit-learn
scipy
tiktoken
tenacity
tqdm

types-PyYAML
types-requests